# server/api/routes/auth.go
package routes

import (
	"net/http"
	"net/mail"
	"strings"
	"unicode/utf8"

	"cgl/api/httpx"
	"cgl/constants"
	"cgl/db"
	"cgl/log"
	"cgl/obj"
)

// RegisterRequest is the request body for user registration
type RegisterRequest struct {
	Name     string `json:"name"`
	Email    string `json:"email"`
	AgeGroup string `json:"ageGroup"` // "u13" (13-17) or "u18" (18+)
}

// RegisterUser godoc
//
//	@Summary		Register new user
//	@Description	Registers a new user after Auth0 authentication. Requires valid Auth0 token.
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			request	body		RegisterRequest	true	"Registration details"
//	@Success		200		{object}	obj.User
//	@Failure		400		{object}	httpx.ErrorResponse	"Invalid request or validation error"
//	@Failure		401		{object}	httpx.ErrorResponse	"Unauthorized"
//	@Failure		409		{object}	httpx.ErrorResponse	"Name already taken"
//	@Security		BearerAuth
//	@Router			/auth/register [post]
func RegisterUser(w http.ResponseWriter, r *http.Request) {
	// Get Auth0 ID from context (set by RequireAuth0Token middleware)
	auth0ID := httpx.Auth0IDFromContext(r.Context())
	if auth0ID == "" {
		httpx.WriteError(w, http.StatusUnauthorized, "Auth0 ID not found in token")
		return
	}

	var req RegisterRequest
	if err := httpx.ReadJSON(r, &req); err != nil {
		httpx.WriteError(w, http.StatusBadRequest, "Invalid JSON: "+err.Error())
		return
	}

	// Validate name
	name := strings.TrimSpace(req.Name)
	if name == "" {
		httpx.WriteError(w, http.StatusBadRequest, "Name is required")
		return
	}
	if utf8.RuneCountInString(name) > 24 {
		httpx.WriteError(w, http.StatusBadRequest, "Name must be at most 24 characters")
		return
	}

	// Check for profanity
	if constants.IsProfane(name) {
		httpx.WriteAppError(w, obj.ErrProfaneName("This name is not allowed"))
		return
	}

	// Validate email
	email := strings.TrimSpace(req.Email)
	if email == "" {
		httpx.WriteError(w, http.StatusBadRequest, "Email is required")
		return
	}
	if _, err := mail.ParseAddress(email); err != nil {
		httpx.WriteError(w, http.StatusBadRequest, "Invalid email address")
		return
	}

	// Validate age group
	ageGroup := strings.TrimSpace(req.AgeGroup)
	if ageGroup == "" {
		httpx.WriteError(w, http.StatusBadRequest, "Age group is required")
		return
	}
	if ageGroup != "u13" && ageGroup != "u18" {
		httpx.WriteError(w, http.StatusBadRequest, "Age group must be 'u13' or 'u18'")
		return
	}

	// Check if user already exists with this Auth0 ID
	log.Debug("checking if user exists by auth0_id", "auth0_id", auth0ID, "email", email, "name", name)
	existingUser, err := db.GetUserByAuth0ID(r.Context(), auth0ID)
	if existingUser != nil {
		log.Warn("user already registered with this auth0_id", "auth0_id", auth0ID, "user_id", existingUser.ID, "existing_email", existingUser.Email)
		httpx.WriteError(w, http.StatusBadRequest, "User already registered")
		return
	}
	if err != nil {
		log.Debug("no user found by auth0_id (expected for new registration)", "auth0_id", auth0ID, "error", err)
	}

	// Also check if a user exists with this email (might be a duplicate or auth0_id mismatch)
	existingByEmail, emailErr := db.GetUserByEmail(r.Context(), email)
	if emailErr == nil && existingByEmail != nil {
		log.Error("user with this email already exists",
			"email", email,
			"existing_user_id", existingByEmail.ID,
			"existing_auth0_id", existingByEmail.Auth0Id,
			"new_auth0_id", auth0ID,
			"auth0_ids_match", existingByEmail.Auth0Id != nil && *existingByEmail.Auth0Id == auth0ID)
	}

	// Check if name is already taken
	nameTaken, err := db.IsNameTaken(r.Context(), name)
	if err != nil {
		log.Error("failed to check name availability", "error", err)
		httpx.WriteError(w, http.StatusInternalServerError, "Failed to check name availability")
		return
	}
	if nameTaken {
		httpx.WriteError(w, http.StatusConflict, "Name is already taken")
		return
	}

	// Create the user
	log.Info("attempting to create user", "name", name, "email", email, "auth0_id", auth0ID)
	user, err := db.CreateUser(r.Context(), name, &email, auth0ID, &ageGroup)
	if err != nil {
		log.Error("failed to create user", "name", name, "email", email, "auth0_id", auth0ID, "error", err)
		httpx.WriteError(w, http.StatusInternalServerError, "Failed to create user")
		return
	}

	log.Info("user registered", "user_id", user.ID, "name", name, "auth0_id", auth0ID, "email", email)
	httpx.WriteJSON(w, http.StatusOK, user)
}

// CheckNameAvailability godoc
//
//	@Summary		Check if name is available
//	@Description	Checks if a username is available for registration
//	@Tags			auth
//	@Produce		json
//	@Param			name	query		string	true	"Name to check"
//	@Success		200		{object}	map[string]bool
//	@Failure		400		{object}	httpx.ErrorResponse	"Invalid request"
//	@Router			/auth/check-name [get]
func CheckNameAvailability(w http.ResponseWriter, r *http.Request) {
	name := strings.TrimSpace(r.URL.Query().Get("name"))
	if name == "" {
		httpx.WriteError(w, http.StatusBadRequest, "Name parameter is required")
		return
	}

	if utf8.RuneCountInString(name) > 24 {
		httpx.WriteJSON(w, http.StatusOK, map[string]bool{"available": false})
		return
	}

	// Check for profanity
	if constants.IsProfane(name) {
		httpx.WriteJSON(w, http.StatusOK, map[string]bool{"available": false, "profane": true})
		return
	}

	taken, err := db.IsNameTaken(r.Context(), name)
	if err != nil {
		log.Error("failed to check name availability", "error", err)
		httpx.WriteError(w, http.StatusInternalServerError, "Failed to check name availability")
		return
	}

	httpx.WriteJSON(w, http.StatusOK, map[string]bool{"available": !taken})
}

// Logout godoc
//
//	@Summary		Logout user
//	@Description	Clears the session cookie for participant users
//	@Tags			auth
//	@Produce		json
//	@Success		200	{object}	map[string]string
//	@Router			/auth/logout [post]
func Logout(w http.ResponseWriter, r *http.Request) {
	httpx.ClearSessionCookie(w)
	log.Debug("session cookie cleared")
	httpx.WriteJSON(w, http.StatusOK, map[string]string{"message": "Logged out"})
}

// ParticipantLoginRequest is the request body for participant token login
type ParticipantLoginRequest struct {
	Token string `json:"token"`
}

// ParticipantLogin godoc
//
//	@Summary		Login with participant token
//	@Description	Logs in a participant using their access token and sets a session cookie
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			request	body		ParticipantLoginRequest	true	"Participant token"
//	@Success		200		{object}	map[string]string
//	@Failure		400		{object}	httpx.ErrorResponse	"Invalid request"
//	@Failure		401		{object}	httpx.ErrorResponse	"Invalid token"
//	@Router			/auth/participant-login [post]
func ParticipantLogin(w http.ResponseWriter, r *http.Request) {
	var req ParticipantLoginRequest
	if err := httpx.ReadJSON(r, &req); err != nil {
		httpx.WriteError(w, http.StatusBadRequest, "Invalid JSON: "+err.Error())
		return
	}

	if req.Token == "" {
		httpx.WriteError(w, http.StatusBadRequest, "Token is required")
		return
	}

	// Validate the token by looking up the user
	user, err := db.GetUserByParticipantToken(r.Context(), req.Token)
	if err != nil {
		log.Debug("participant login failed: invalid token", "error", err)
		httpx.WriteError(w, http.StatusUnauthorized, "Invalid or expired token")
		return
	}

	// Set the session cookie with the participant token
	httpx.SetSessionCookie(w, r, req.Token)

	log.Info("participant login successful", "user_id", user.ID, "name", user.Name)
	httpx.WriteJSON(w, http.StatusOK, map[string]string{"message": "Logged in", "userId": user.ID.String()})
}