# server/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
	"encoding/json"
	"time"

	"github.com/google/uuid"
	"github.com/sqlc-dev/pqtype"
)

type ApiKey struct {
	ID               uuid.UUID
	CreatedBy        uuid.NullUUID
	CreatedAt        time.Time
	ModifiedBy       uuid.NullUUID
	ModifiedAt       time.Time
	UserID           uuid.UUID
	Name             string
	Platform         string
	Key              string
	IsDefault        bool
	LastUsageSuccess sql.NullBool
	LastErrorCode    sql.NullString
}

type ApiKeyShare struct {
	ID            uuid.UUID
	CreatedBy     uuid.NullUUID
	CreatedAt     time.Time
	ModifiedBy    uuid.NullUUID
	ModifiedAt    time.Time
	ApiKeyID      uuid.UUID
	UserID        uuid.NullUUID
	WorkshopID    uuid.NullUUID
	InstitutionID uuid.NullUUID
	GameID        uuid.NullUUID
}

type AppUser struct {
	ID                   uuid.UUID
	CreatedBy            uuid.NullUUID
	CreatedAt            time.Time
	ModifiedBy           uuid.NullUUID
	ModifiedAt           time.Time
	Name                 string
	Email                sql.NullString
	DeletedAt            sql.NullTime
	Auth0ID              sql.NullString
	ParticipantToken     sql.NullString
	DefaultApiKeyShareID uuid.NullUUID
	AiQualityTier        sql.NullString
	Language             string
	PrivateShareID       uuid.NullUUID
	AgeGroup             sql.NullString
}

type Game struct {
	ID                           uuid.UUID
	CreatedBy                    uuid.NullUUID
	CreatedAt                    time.Time
	ModifiedBy                   uuid.NullUUID
	ModifiedAt                   time.Time
	Name                         string
	Description                  string
	Icon                         []byte
	WorkshopID                   uuid.NullUUID
	Public                       bool
	PublicSponsoredApiKeyShareID uuid.NullUUID
	SystemMessageScenario        string
	SystemMessageGameStart       string
	ImageStyle                   string
	Css                          string
	StatusFields                 string
	Theme                        pqtype.NullRawMessage
	FirstMessage                 sql.NullString
	FirstStatus                  sql.NullString
	FirstImage                   []byte
	OriginallyCreatedBy          uuid.NullUUID
	PlayCount                    int32
	CloneCount                   int32
	DeletedAt                    sql.NullTime
}

type GameSession struct {
	ID                       uuid.UUID
	CreatedBy                uuid.NullUUID
	CreatedAt                time.Time
	ModifiedBy               uuid.NullUUID
	ModifiedAt               time.Time
	GameID                   uuid.UUID
	UserID                   uuid.UUID
	WorkshopID               uuid.NullUUID
	ApiKeyID                 uuid.NullUUID
	AiPlatform               string
	AiModel                  string
	AiSession                json.RawMessage
	ImageStyle               string
	Language                 string
	StatusFields             string
	Theme                    pqtype.NullRawMessage
	IsOrganisationUnverified bool
	DeletedAt                sql.NullTime
}

type GameSessionMessage struct {
	ID                     uuid.UUID
	CreatedBy              uuid.NullUUID
	CreatedAt              time.Time
	ModifiedBy             uuid.NullUUID
	ModifiedAt             time.Time
	GameSessionID          uuid.UUID
	Seq                    int32
	Type                   string
	Message                string
	Status                 sql.NullString
	Plot                   sql.NullString
	ImagePrompt            sql.NullString
	Image                  []byte
	Audio                  []byte
	HasImage               bool
	HasAudio               bool
	PromptStatusUpdate     sql.NullString
	PromptResponseSchema   sql.NullString
	PromptImageGeneration  sql.NullString
	PromptExpandStory      sql.NullString
	ResponseRaw            sql.NullString
	TokenUsage             pqtype.NullRawMessage
	UrlAnalytics           sql.NullString
	ApiKeyType             sql.NullString
	PromptConstraintSource sql.NullString
	DeletedAt              sql.NullTime
}

type GameShare struct {
	ID            uuid.UUID
	GameID        uuid.UUID
	Token         string
	ApiKeyShareID uuid.UUID
	InstitutionID uuid.NullUUID
	WorkshopID    uuid.NullUUID
	Remaining     sql.NullInt32
	AiQualityTier sql.NullString
	CreatedBy     uuid.NullUUID
	CreatedAt     time.Time
}

type GameTag struct {
	ID         uuid.UUID
	CreatedBy  uuid.NullUUID
	CreatedAt  time.Time
	ModifiedBy uuid.NullUUID
	ModifiedAt time.Time
	GameID     uuid.UUID
	Tag        string
}

type Institution struct {
	ID                   uuid.UUID
	CreatedBy            uuid.NullUUID
	CreatedAt            time.Time
	ModifiedBy           uuid.NullUUID
	ModifiedAt           time.Time
	Name                 string
	DeletedAt            sql.NullTime
	FreeUseApiKeyShareID uuid.NullUUID
	FreeUseAiQualityTier sql.NullString
	PromptConstraints    sql.NullString
}

type SystemSetting struct {
	ID                   uuid.UUID
	CreatedAt            time.Time
	ModifiedAt           time.Time
	DefaultAiQualityTier string
	FreeUseAiQualityTier sql.NullString
	PromptConstraintU13  sql.NullString
	PromptConstraintU18  sql.NullString
	SchemaVersion        int32
	FreeUseApiKeyID      uuid.NullUUID
}

type UserFavouriteGame struct {
	ID         uuid.UUID
	CreatedBy  uuid.NullUUID
	CreatedAt  time.Time
	ModifiedBy uuid.NullUUID
	ModifiedAt time.Time
	UserID     uuid.UUID
	GameID     uuid.UUID
}

type UserRole struct {
	ID               uuid.NullUUID
	CreatedBy        uuid.NullUUID
	CreatedAt        time.Time
	ModifiedBy       uuid.NullUUID
	ModifiedAt       time.Time
	UserID           uuid.UUID
	Role             sql.NullString
	InstitutionID    uuid.NullUUID
	WorkshopID       uuid.NullUUID
	ActiveWorkshopID uuid.NullUUID
}

type UserRoleInvite struct {
	ID            uuid.UUID
	CreatedBy     uuid.NullUUID
	CreatedAt     time.Time
	ModifiedBy    uuid.NullUUID
	ModifiedAt    time.Time
	InstitutionID uuid.UUID
	Role          string
	WorkshopID    uuid.NullUUID
	InvitedUserID uuid.NullUUID
	InvitedEmail  sql.NullString
	InviteToken   sql.NullString
	MaxUses       sql.NullInt32
	UsesCount     int32
	ExpiresAt     sql.NullTime
	Status        string
	DeletedAt     sql.NullTime
	AcceptedAt    sql.NullTime
	AcceptedBy    uuid.NullUUID
}

type Workshop struct {
	ID                         uuid.UUID
	CreatedBy                  uuid.NullUUID
	CreatedAt                  time.Time
	ModifiedBy                 uuid.NullUUID
	ModifiedAt                 time.Time
	Name                       string
	InstitutionID              uuid.UUID
	Active                     bool
	Public                     bool
	DeletedAt                  sql.NullTime
	DefaultApiKeyShareID       uuid.NullUUID
	AiQualityTier              sql.NullString
	PromptConstraints          sql.NullString
	ShowPublicGames            bool
	ShowOtherParticipantsGames bool
	DesignEditingEnabled       bool
	IsPaused                   bool
	AllowGameSharing           bool
}

type WorkshopParticipant struct {
	ID          uuid.UUID
	CreatedBy   uuid.NullUUID
	CreatedAt   time.Time
	ModifiedBy  uuid.NullUUID
	ModifiedAt  time.Time
	WorkshopID  uuid.UUID
	Name        string
	AccessToken string
	Active      bool
}