# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for ChatGameLab - an AI-powered interactive game platform",
        "title": "ChatGameLab API",
        "contact": {
            "name": "ChatGameLab Team",
            "url": "https://chatgamelab.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/auth/check-name": {
            "get": {
                "description": "Checks if a username is available for registration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Check if name is available",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name to check",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/participant-login": {
            "post": {
                "description": "Logs in a participant using their access token and sets a session cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login with participant token",
                "parameters": [
                    {
                        "description": "Participant token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.ParticipantLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a new user after Auth0 authentication. Requires valid Auth0 token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "Registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/obj.User"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Name already taken",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                },
                "type": {
                    "description": "Deprecated: use Code instead",
                    "type": "string"
                }
            }
        },
...
        "obj.Institution": {
            "type": "object",
            "properties": {
                "freeUseAiQualityTier": {
                    "description": "high/medium/low, nil = server default",
                    "type": "string"
                },
                "freeUseApiKeyShareId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/obj.InstitutionMember"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/obj.Meta"
                },
                "name": {
                    "type": "string"
                },
                "promptConstraints": {
                    "type": "string"
                }
            }
        },
        "obj.InstitutionMember": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/obj.Role"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "obj.InviteStatus": {
            "type": "string",
            "enum": [
                "pending",
                "accepted",
                "declined",
                "expired",
                "revoked"
            ],
            "x-enum-varnames": [
                "InviteStatusPending",
                "InviteStatusAccepted",
                "InviteStatusDeclined",
                "InviteStatusExpired",
                "InviteStatusRevoked"
            ]
        },
        "obj.Meta": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "$ref": "#/definitions/uuid.NullUUID"
                },
                "modifiedAt": {
                    "type": "string"
                },
                "modifiedBy": {
                    "$ref": "#/definitions/uuid.NullUUID"
                }
            }
        },
        "obj.Role": {
            "type": "string",
            "enum": [
                "admin",
                "head",
                "staff",
                "participant",
                "individual"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleHead",
                "RoleStaff",
                "RoleParticipant",
                "RoleIndividual"
            ]
        },
...
        "obj.User": {
            "type": "object",
            "properties": {
                "ageGroup": {
                    "description": "\"u13\" (13-17) or \"u18\" (18+), nil for guests",
                    "type": "string"
                },
                "aiQualityTier": {
                    "description": "high/medium/low, nil = server default",
                    "type": "string"
                },
                "auth0Id": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "description": "ISO 639-1 language code (e.g., \"en\", \"de\", \"fr\")",
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/obj.Meta"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/obj.UserRole"
                }
            }
        },
        "obj.UserRole": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "institution": {
                    "$ref": "#/definitions/obj.Institution"
                },
                "meta": {
                    "$ref": "#/definitions/obj.Meta"
                },
                "role": {
                    "$ref": "#/definitions/obj.Role"
                },
                "userId": {
                    "type": "string"
                },
                "workshop": {
                    "$ref": "#/definitions/obj.Workshop"
                }
            }
        },
        "obj.UserRoleInvite": {
            "type": "object",
            "properties": {
                "acceptedAt": {
                    "type": "string"
                },
                "acceptedBy": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "institutionId": {
                    "type": "string"
                },
                "inviteToken": {
                    "type": "string"
                },
                "invitedEmail": {
                    "type": "string"
                },
                "invitedUserId": {
                    "type": "string"
                },
                "maxUses": {
                    "type": "integer"
                },
                "meta": {
                    "$ref": "#/definitions/obj.Meta"
                },
                "role": {
                    "$ref": "#/definitions/obj.Role"
                },
                "status": {
                    "$ref": "#/definitions/obj.InviteStatus"
                },
                "usesCount": {
                    "type": "integer"
                },
                "workshopId": {
                    "type": "string"
                }
            }
        },
...
        "obj.Workshop": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "aiQualityTier": {
                    "description": "Workshop settings (configured by staff/heads)",
                    "type": "string"
                },
                "allowGameSharing": {
                    "type": "boolean"
                },
                "defaultApiKeyShareId": {
                    "type": "string"
                },
                "designEditingEnabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "institution": {
                    "$ref": "#/definitions/obj.Institution"
                },
                "invites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/obj.UserRoleInvite"
                    }
                },
                "isPaused": {
                    "type": "boolean"
                },
                "meta": {
                    "$ref": "#/definitions/obj.Meta"
                },
                "name": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/obj.WorkshopParticipant"
                    }
                },
                "promptConstraints": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "showOtherParticipantsGames": {
                    "type": "boolean"
                },
                "showPublicGames": {
                    "type": "boolean"
                }
            }
        },
        "obj.WorkshopParticipant": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "active": {
                    "type": "boolean"
                },
                "gamesCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/obj.Meta"
                },
                "name": {
                    "type": "string"
                },
                "permanent": {
                    "type": "boolean"
                },
                "role": {
                    "$ref": "#/definitions/obj.Role"
                },
                "workshopId": {
                    "type": "string"
                }
            }
        },
...
        "routes.ParticipantLoginRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "routes.RegisterRequest": {
            "type": "object",
            "properties": {
                "ageGroup": {
                    "description": "\"u13\" (13-17) or \"u18\" (18+)",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if UUID is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token in format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}